add illuminate orm composer plugin.
